+++ ParallaxGen/src/main.cpp
@@ -1,6 +1,7 @@
 #include <CLI/CLI.hpp>
 
 #include <NifFile.hpp>
+#include <boost/algorithm/string/predicate.hpp>
 #include <spdlog/common.h>
 #include <spdlog/logger.h>
 #include <spdlog/sinks/rotating_file_sink.h>
@@ -194,14 +195,9 @@ void mainRunner(ParallaxGenCLIArgs &Args, const filesystem::path &ExePath) {
     exit(1);
   }
 
-  // delete existing output
-  PG.deleteOutputDir();
-
-  // Check if ParallaxGen output already exists in data directory
-  const filesystem::path PGStateFilePath = BG.getGameDataPath() / ParallaxGen::getDiffJSONName();
-  if (filesystem::exists(PGStateFilePath)) {
-    spdlog::critical("ParallaxGen meshes exist in your data directory, please delete before "
-                     "re-running.");
+  // If output dir is a subdirectory of data dir vfs issues can occur
+  if (boost::istarts_with(Params.Output.Dir.wstring(), BG.getGameDataPath().wstring() + "\\")) {
+    spdlog::critical("Output directory cannot be a subdirectory of your data folder. Exiting.");
     exit(1);
   }
 
@@ -217,12 +213,32 @@ void mainRunner(ParallaxGenCLIArgs &Args, const filesystem::path &ExePath) {
   if (Params.ModManager.Type == ModManagerDirectory::ModManagerType::ModOrganizer2 &&
       !Params.ModManager.MO2InstanceDir.empty() && !Params.ModManager.MO2Profile.empty()) {
     // MO2
-    MMD.populateModFileMapMO2(Params.ModManager.MO2InstanceDir, Params.ModManager.MO2Profile);
+    MMD.populateModFileMapMO2(Params.ModManager.MO2InstanceDir, Params.ModManager.MO2Profile, Params.Output.Dir);
   } else if (Params.ModManager.Type == ModManagerDirectory::ModManagerType::Vortex) {
     // Vortex
     MMD.populateModFileMapVortex(BG.getGameDataPath());
   }
 
+  // delete existing output
+  PG.deleteOutputDir();
+
+  // Check if ParallaxGen output already exists in data directory
+  const filesystem::path PGStateFilePath = BG.getGameDataPath() / ParallaxGen::getDiffJSONName();
+  if (filesystem::exists(PGStateFilePath)) {
+    spdlog::critical("ParallaxGen meshes exist in your data directory, please delete before "
+                     "re-running.");
+    exit(1);
+  }
+
+  // Check if dyndolod.esp exists
+  const auto ActivePlugins = BG.getActivePlugins(false, true);
+  if (find(ActivePlugins.begin(), ActivePlugins.end(), L"dyndolod.esp") != ActivePlugins.end()) {
+    spdlog::critical("DynDoLOD and TexGen outputs must be disabled prior to running ParallaxGen. It is recommended to "
+                     "generate LODs after running ParallaxGen with the ParallaxGen output enabled.");
+    exit(1);
+  }
+
+  // Init file map
   PGD.populateFileMap(Params.Processing.BSA);
 
   // Map files